Include a copy of tempest.scenario.manager module
Add tempest/scenario/manager.py from openstack/tempest [1], and update
imports in ironic. As announced in [2], the interface in manager.py is
not stable, and should not be consumed outside tempest. With this
temporary change, QA will have time to improve the tempest.scenario
interface for other projects to consume, without breaking gate jobs.

An alternative to copying the local file was proposed in [3], which
suggested creating a new interface class, living along with the old
interface. Through a tempest config option projects would choose which
scenario interface class to use. QA reasoned that this would increase
the number of jobs, and slow their work. Thus the option was rejected.

[1] http://git.openstack.org/cgit/openstack/tempest/tree/tempest/scenario/manager.py?id=82a278e88c9e9f9ba49f81c1f8dba0bca7943daf
[2] http://lists.openstack.org/pipermail/openstack-dev/2017-February/112938.html
[3] http://lists.openstack.org/pipermail/openstack-dev/2017-March/113496.html

Closes-Bug: #1668807
Change-Id: Iaa06df6ab75fb8f79cbcd3f6b21e68622e7799c1