Use the main introspection test instead of the smoke one
Currently we have two introspection tests:
- Full, running in the regular job. Deploys after introspection.
- Smoke, running in grenade. Only does introspection.

This change leaves only the first test and marks at as smoke,
so that it runs in grenade. This way we get better upgrade testing
coverage at the expense of slightly longer grenade run time.

This change makes it unnecessary to run inspector grenade jobs
on ironic-tempest-plugin, so they are removed.

Change-Id: Ie28a1d2122d3ea49fc3d39864e3f149aae4c1190