Explicitly use iscsi for standalone BaremetalIpmiRescuePartitioned
This test uses partition images and netboot, which the ansible deploy
interface does not support. Since we only update the interfaces we
specify when choosing a node, if we choose a node that's already using
the ansible interface, we'll fail here. Explicitly set the iscsi
interface to avoid this situation.

Depends-On: https://review.openstack.org/563119
Depends-On: https://review.openstack.org/563120
Change-Id: I852fbe8d2d34a5d301d3591af6cd67c4dcda3b83