[gate] update the list of the voting jobs
Make jobs running with N-3 (currently Pike) and older non-voting
(and thus remove them from the gate queue). I have a gut feeling
that a change that breaks N-3 is very likely to break N-2 (currently
Queens) as well, so it's enough to have N-2 voting.

Make the discovery and the multinode jobs from all stable branches
non-voting.  These jobs cover the tests that get changed very infrequently
(if ever).  These are also the jobs with the highest random failure rate.

Add missing API tests for Rocky (this job did not exist before that).
Use the updated job name for the multinode job.

Change-Id: Icf97a9cd5b88d2263551f32c0d3c1e09a712b346