Deploy templates: add API tests
Adds tests for the deploy templates API added in 1.55.

Also fixes an issue where the microversion fixture was not used in the resource
cleanup, so new resources would see a 404 and not be deleted. Fixing this
uncovered an issue in the volume tests where volume connectors and targets
could not be deleted due to being associated with a powered on node. The
simplest fix was to move node deletion before the volume connector and target
in resource cleanup.

Change-Id: I09d9a21e1ea5512c0140e818b0ca3de501870c12
Depends-On: https://review.openstack.org/631845
Story: 1722275
Task: 28679