Temporary marking ironic-standalone non-voting
The ironic-standalone job is failing because the most recent IPA
image has only python3 and ansible uses python 2 as default
interpreter, resulting in a /usr/bin/python not found error.
Marking ironic-standalone non voting and removing it from the
gate until we can run ansible with python3 on remote machines.

Change-Id: Id03b02ea805809c35c4ef34bd1b255d7f6e6d4a7