Enable boot_option to be specified
The ability to set this is required as some test cases do not
support a duality of local boot and network boot when using
the cirros test image. This is because cirros lacks the tools
to install a grub2 boot loader.

Adds a ``[baremetal]partition_netboot`` configuration parameter
which defaults to ``True``. This allows a test executor to tell
the tests if we should expect to network boot all partition
image usage. This is ultimately required to enable ironic to
change the underlying default boot_option netboot to local.

Story: #1619339

Change-Id: I784d5c9d98d574fc1f00dd8f5ee1e20499bf2e34