Make local boot explicit on software raid tests
Software RAID requies the installation of a boot loader
and in this scenario, we need to explicitly state that
local boot is desired since ironic deployments do have
a default_boot_option which can be set.

In the default case of netboot, these tests will silently
work if the content written to disk is also bootable
because the netboot default causes ironic to skip
bootloader installation.

Change-Id: I583840f2a50753fd169149d76c7123392eb362b5
Task: 36415
Story: 2006474