Remove /extra/vif_port_id usage
A long time ago, in a galaxy far far away, the use of
/extra/vif_port_id was deprecated in favor of vif attachment and
detachment. This change updates a test which was using the
/extra/vif_port_id field value to bind a vif, and then list/detach
it using the vif attach/detach/list API interfaces.

Since the action is to add/list/detach, this seems appropriate and
should be widely supported going back 4+ years.

Change-Id: I0b0cd0aaf9a87803386e096ab768638476f9ca52