Fix for verifying introspected data
This fixes verification of introspected data.

For physical baremetal introspection scenario, specification of the
physical server may differ with existing flavor data, whereas for
virtualised baremetal server, it can be matched.

When the introspection is performed on physical baremetal server,
the introspected data does not match with flavor data causing failure
in verification of introspected data. Hence, current fix is verifying
introspected data if greater than 0.

[1] https://docs.openstack.org/ironic/latest/install/configure-nova-flavors.html

Change-Id: I9ff1364e1baf99579a3d7b8cb5fc2ab012030173