Undo maintenace state on protected node tests
As part of the protected from deletion test, we set a node into maintenance
state which allows override actions to be performed for node deletions.

The test is verifying the removal of the node is prevented. This fails upon
the TearDown class execution, resulting in a silent hidden failure which
cannot complete teardown because you cannot unprovision, as is in the
teardown step, a node in maintenance state.

Explicitly unset maintenance, and the world will be happy!

Change-Id: I89839ecb0e08555ddcfee0289350fac1923b3958