Disable use of the SSH agent for the rescue test
Paramiko, by default, attempts agent based authentication before
ever attempting password based authentication. But this doesn't
work so well if the environment has an agent running or thinks it
has an agent running, because it breaks the rescue test.

Instead, explicitly tell the tempest and thus paramiko to *not* use
the ssh-agent functionality.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/872566
Change-Id: Ia2870579c7398d359b8d360fe7d1e42abb79b299