Add test for dhcp-less vmedia based deployment
Creating test to help facilitate the fix of bug 2032377 and
ultimately help ensure we have a backwards compatible fix.

The scenario:

* Creates a shiny new network without dhcp, and adds a router.
* Sets that network as the provisioning network for the node.
* Creates a port on that network, assigns that port as the VIF.
* Creates a configuration drive network_data.json file payload
* Triggers deployment, utilizing the aformentioned network_data
  payload and configuration drive.
* Once deployment has complete, attempts to ping the remote host.
* Once pingable, and this is configurable, it will then attempt
  to rebuild the remote host, basically forcing the condition
  covered in bug #2032377.
* Rebuild is completed, and the host is attempted to be pinged,
  again.

To facilitate this, three configuration options have been added.
Two are external network settings. Because there is no way in
a devstack configuration of neutron to advertise the next hop
router, we have to have a pre-assigned/configured IP on the
external network we can attach a router to. We also need to know
the external network ID.

Then there is basically a option flag if we wish to exercise the
pattern for bug 2032377. Ideally, this would be always, but with
the complexity and fact a non-stock IPA image is required, one
sort of needs to know and then configure as appropriate.

Change-Id: Ic848b8051e4d863f30d47c833d334afc879e4f20