Test multiple boot interfaces as part of one CI job
The idea here came out of the 2024.1 PTG held in ?October? 2023,
with the goal of being able to exercise testing of boot interfaces
as a group in a single job, instead of spreading it out amongst
CI jobs with different scenarios.

That doesn't mean different scenarios are wrong, but as we add more
boot interfaces as part of supporting HTTPBoot, we cannot double
our scenarios.

In theory, the hope is we should be able to consolidate down several
secenario jobs into just one test running these jobs.

Change-Id: Id41ee40c36624387216b2cfc1501f01a3217b2af