Invoke tests with fake interfaces
So it turns out as highlighted by looking at bug 2054722
that you can end up creating nodes with "fake" hardware nodes
in the default config we test in the gate, which end up with
"fake" as the deploy_interface, but if you try to run the
same test against a production configured ironic deployment,
it fails because it likely ends up with a default of "agent"
as the deploy_interface.

Unfortunately, a review of test logs also reveals that the
same basic problem exists with the network_interface field
*as well*, where tests were written expecting the default
to be a fake or noop interface out of the box.

Note: This depends-on is to be removed before mering, it is
      enable additional issues to be identified.

Change-Id: Id0051d9b39bc0f46e3afee5bbfa8a2062114df80