Patch to enforce json extension works in existing API behaviour
Currently, .json extensions are stripped out during the search or while
taking any action on a node through its name or uuid.

This pecan feature ensures that a reference to
``0178-0c2c-9c26-ca69-3011-a9dd.json`` is treated
as ``0178-0c2c-9c26-ca69-3011-a9dd``

If we had two nodes that differ by a .json extension in either of the
names, ``"node", "node.json"``, and we intend for the latter, we’ll
need to append an additional ".json" extension in the reference query,
else the node without the extension is treated as the default node,
and the extension node is only considered if it’s the only node with
the name.

Same workaround if we wanted our ``0178-0c2c-9c26-ca69-3011-a9dd.json``
correctly targeted, we append an additonal ".json" extension to our
query.

This patch tries to validate this behavior in API versions prior to
1.91, where this feature exists.

Change-Id: I613e596e8cf52f36b9c7caac30bdb3ff0d3bf821