Delete the CUSTOM_RAID template after test has finished
Previously, the tempest plugin would orphan a CUSTOM_RAID
deploy template in the ironic deployment by not removing it
after the test has been completed. This change adds a cleanup
to trigger the deletion of the template, so the test does not
fail upon a second re-execution against the environment.

Closes-Bug: 2055456
Change-Id: I21ea329feb48e56b7f2984f62cdbc41c1a0bd1d0