Vmedia advanced ops: Clarify config drive and don't ping after rebuild
Turns out there are issues if your rebuild.

The prime issue is if the config drive is stored in swift. If that
temporary URL expires, then a rebuild fails.

As such, and since there is a relatvely low value in pinging the host,
as the test is about ensuring ironic and the agent are working in
concert with virtual media, we just wait until the appropriate state
is reached and then begin to exit the test as we can be confident
that communication is working as expected.

Change-Id: I64a805b0425687268f36b858304981b0b1fc09db