Handle invalid configuration for vif test execution
It is entirely possible to run the tempest tests without a physical
network parameter for the compute service. Turns out the network is
required when you use [network]shared_physical_network set to true,
because without it we can't figure out how to attach to the network.

So now we skip when the input configuration is such that there is
no way for the tests to proceed.

Closes-Bug: 2065378
Change-Id: Ic448a1a6383c2da2567b9da2262d4572536296ed