Exclude ramdisk tests with tinycore in uefi mode
Tinycore artifacts are not uefi bootable unless the firmware
jumps back to bios booting the node. This presents with ramdisk
tests when a node may end up using the tinycore linux iso image if
a boot iso cannot be reliably built in Ironic's CI due to CI resource
provider constraints.

Instead, we now will just skip the test if we detect this case.

Closes-Bug: 2065811
Change-Id: Iae0ba6a52aa81f02202da768d4f755469badfe33