Work around missing microversion headers on /
In get_min_max_api_microversions, we issue a request against the root
endpoint / to get the available microversions. This endpoint does return
them in the body but due to an issue in Ironic does not return the
microversion headers. Unfortunately, our implementation of request()
makes Tempest expected these headers and fail if they're absent.

Until Ironic is fixed, use /v1 to get the required information.

Change-Id: I7003e0c90ce764c903d870d739786a8a97d5f0af
Related-Bug: #2079023