CI: scale back the number of inspector and scope enforced jobs
* Remove specific scope enforced jobs from master

With recent oslo.policy changes, we no longer need to have specific
scope enforced test jobs. As such, removing them from master branch
testing.

* Leave only the master ironic-inspector job voting

Inspector is deprecated and does not receive many changes. Leave one
voting jobs and 2 non-voting: for 2024.1 and the discovery one.

* Remove inspector scope enforced jobs from the gate

They are not in the check pipeline and must not be in the gate.
Same logic as above applies to these jobs.

* Copy anaconda jobs into the gate pipeline

Voting check jobs must be in the gate, so fixing the descrepancy.

* Purge definitions of unused jobs.

Co-Authored-By: Dmitry Tantsur <dtantsur@protonmail.com>
Change-Id: I17e1604b036e98ee2e8ce6424bfeb2b0bef8e471