Fix failure cases around test_ports
In some cases where operators seek to use ironic-tempest-plugin,
there may not be complete networking configuration behind ironic,
like site specific cleaning and provisioning networks, which is
fine, that is a choice the operator has made, but then if the
default network interface, which these tests were modeled upon
is a different network interface, such as flat or neutron,
then these tests will fail.

So instead of rely upon the state configuration of ironic
in terms of what network interface to end up with, we now assert
the noop network interface to explicitly tell ironic "we know what
we want/expect" for the test to operate.

This sort of class of failure has come up a few times with slight
variations, where functionally the test suite just was developed
expecting everything under the backend driver hood to be noop.

Closes-Bug: 2087263
Change-Id: Iea732f58c23dd0c79305a86b6707fb750aabf30f