Remove reboot from basic baremetal ops
While investigating the 4k job, I noticed the test sometimes has trouble
SSHing back in to the host post-reboot due to no fault of the test.

In other tests, it looks like we manage to get SSH'ed in just as the
deployed host is online.

This is because cloud-init or other configuration changes have been
applied in the OS, which is outside the act of "deploying" the node.

In other words, we were spinning our basic job for several hundred
seconds in some cases on just waiting for the OS to reboot, and then
trying to verify the OS behaves the same as before we rebooted.

Ultimately, to conserve resources *and* reduce failure risk by trying
to test OS behavior, removing the reboot call and post-reboot re-ssh/ping
check behavior.

Change-Id: Ie08335e0e10f51f782a27270c62e33aa84a8c568