Replaced all punctuation with an underscore in the node resource class.
If we want to associate a flavor with a custom resource class, it must follow the rules (refer tohttps://docs.openstack.org/ironic/latest/install/configure-nova-flavors.html ). However, if we only want to define a class of nodes, it does not have to follow these rules. The formatting of the node resource class has been adjusted to pass the test.

Change-Id: Ia14cbaa2c53d6a6c3e22f1a10b70597aec4afeca