Remove legacy standalone jobs from test execution
The ironic-standalone job is largely centered around ipmi execution
CI jobs, however IPMI is fairly static and doesn't deliver the overall
project a ton of testing value as compared to Redfish which is the
expected path for most users to use.

As such, removing the ironic-standalone job invocations to allow us
to focus on just running the redfish jobs for tempest.

If we need to make ipmitool specific testing or changes, we can run
those in a more specific context or as part of Ironic's CI jobs
as opposed to the tempest plugin specifically.

Change-Id: I7d45eecaa0075c9b0f489d30e4ca248f26ecf8c8