Add Tempest API tests for inspection rules
As of the 2025.1 "Epoxy" release, inspection rules have been migrated
from the ironic-inspector project into Ironic itself.

This commit introduces Tempest tests to validate the creation, updation
and deletion of new inspection rules functionality within Ironic.

Current focus is on covering the important CRUD operations for
inspection rules to ensure robustness and correctness of
the Inspection rules API.

I guess this is the starting point, as Inspection rules evolves
more tests should be added to ensure the continued relevance
effectiveness of the API test suite.

closes-bug: 2105478

Change-Id: I3f6de02acee8d8c3764d3b1465b92292be3b690c
Signed-off-by: abhibongale <abhishekbongale@outlook.com>