Test graphical console novnc connection
This creates a fake-graphical console and confirms that the websocket
returns the first part of an RFB handshake. For the systemd console
container provider this confirms that the container is running and
listening for incoming VNC connections. This is the same level of test
coverage that nova-novncproxy has in tempest.

Change-Id: Iad118cf2d031125450c00f83b7edfc3f0cedbd1a
Signed-off-by: Steve Baker <sbaker@redhat.com>