ci: fix and log errors on advanced tests disqualification
Apparently we're getting an error in some cases, but also
the base job configuration changed at some point which mooted
the purpose of some of the code. So now, only trigger the excess
code and log an error if there was a failure so it is possible to
understand the actual failure.

Change-Id: I33236532d826e3795497a51bab78d2ceba55285e
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>